refactor(team): extract scheduler into broker_scheduler.go
Slice 12. Moves scheduler-job state management (~353 lines) into a new
broker_scheduler.go:
- Locked accessor: dueSchedulerJobsLocked
- Public lifecycle: SetSchedulerJob, ScheduleTaskFollowUp,
  ScheduleRequestFollowUp, ScheduleRecheck
- Internal: scheduleJob, scheduleJobLocked, completeSchedulerJobsLocked,
  scheduleTaskLifecycleLocked, scheduleRequestLifecycleLocked
- Helpers: normalizeSchedulerSlug, normalizeSlugPart,
  normalizeSchedulerJob, schedulerJobMatches, schedulerJobDue
- HTTP handler: handleScheduler

Relocates TestSchedulerDueOnlyFiltersFutureJobs from broker_tasks_test.go
into a new broker_scheduler_test.go.

Adds three characterization tests:
- TestNormalizeSchedulerSlug_StripsAndJoinsParts (slug composition rule:
  underscore→dash, empty parts dropped, ":"-joined)
- TestSchedulerJobDue_BoundaryAtExactNow (DueAt==now counts as due,
  Status=done/canceled never counts, case-insensitive Status)
- TestCompleteSchedulerJobsLocked_NoOpForUnknownTarget (target mismatch
  must not flip job state)

Total package coverage: 63.1% (vs 62.9% baseline floor).
